[pythonic resources] Improve nested resource evaluation logic
