+++ python_modules/dagster/dagster_tests/core_tests/resource_tests/pythonic_resources/test_general_pythonic_resources.py
@@ -1,8 +1,7 @@
 import enum
-import json
 import sys
 from abc import ABC, abstractmethod
-from typing import Any, Callable, List, Mapping, Optional
+from typing import List, Mapping, Optional
 
 import mock
 import pytest
@@ -14,7 +13,6 @@ from dagster import (
     ConfigurableResource,
     DagsterInstance,
     Definitions,
-    Field,
     IAttachDifferentObjectToOpContext,
     InitResourceContext,
     IOManager,
@@ -424,327 +422,6 @@ def test_runtime_config_run_config_obj():
     assert out_txt == ["greeting: hello, world!"]
 
 
-def test_nested_resources():
-    out_txt = []
-
-    class Writer(ConfigurableResource, ABC):
-        @abstractmethod
-        def output(self, text: str) -> None:
-            pass
-
-    class WriterResource(Writer):
-        def output(self, text: str) -> None:
-            out_txt.append(text)
-
-    class PrefixedWriterResource(Writer):
-        prefix: str
-
-        def output(self, text: str) -> None:
-            out_txt.append(f"{self.prefix}{text}")
-
-    class JsonWriterResource(
-        Writer,
-    ):
-        base_writer: Writer
-        indent: int
-
-        def output(self, obj: Any) -> None:
-            self.base_writer.output(json.dumps(obj, indent=self.indent))
-
-    @asset
-    def hello_world_asset(writer: JsonWriterResource):
-        writer.output({"hello": "world"})
-
-    # Construct a resource that is needed by another resource
-    writer_resource = WriterResource()
-    json_writer_resource = JsonWriterResource(indent=2, base_writer=writer_resource)
-
-    assert (
-        Definitions(
-            assets=[hello_world_asset],
-            resources={
-                "writer": json_writer_resource,
-            },
-        )
-        .get_implicit_global_asset_job_def()
-        .execute_in_process()
-        .success
-    )
-
-    assert out_txt == ['{\n  "hello": "world"\n}']
-
-    # Do it again, with a different nested resource
-    out_txt.clear()
-    prefixed_writer_resource = PrefixedWriterResource(prefix="greeting: ")
-    prefixed_json_writer_resource = JsonWriterResource(
-        indent=2, base_writer=prefixed_writer_resource
-    )
-
-    assert (
-        Definitions(
-            assets=[hello_world_asset],
-            resources={
-                "writer": prefixed_json_writer_resource,
-            },
-        )
-        .get_implicit_global_asset_job_def()
-        .execute_in_process()
-        .success
-    )
-
-    assert out_txt == ['greeting: {\n  "hello": "world"\n}']
-
-
-def test_nested_resources_multiuse():
-    class AWSCredentialsResource(ConfigurableResource):
-        username: str
-        password: str
-
-    class S3Resource(ConfigurableResource):
-        aws_credentials: AWSCredentialsResource
-        bucket_name: str
-
-    class EC2Resource(ConfigurableResource):
-        aws_credentials: AWSCredentialsResource
-
-    completed = {}
-
-    @asset
-    def my_asset(s3: S3Resource, ec2: EC2Resource):
-        assert s3.aws_credentials.username == "foo"
-        assert s3.aws_credentials.password == "bar"
-        assert s3.bucket_name == "my_bucket"
-
-        assert ec2.aws_credentials.username == "foo"
-        assert ec2.aws_credentials.password == "bar"
-
-        completed["yes"] = True
-
-    aws_credentials = AWSCredentialsResource(username="foo", password="bar")
-    defs = Definitions(
-        assets=[my_asset],
-        resources={
-            "s3": S3Resource(bucket_name="my_bucket", aws_credentials=aws_credentials),
-            "ec2": EC2Resource(aws_credentials=aws_credentials),
-        },
-    )
-
-    assert defs.get_implicit_global_asset_job_def().execute_in_process().success
-    assert completed["yes"]
-
-
-def test_nested_resources_runtime_config():
-    class AWSCredentialsResource(ConfigurableResource):
-        username: str
-        password: str
-
-    class S3Resource(ConfigurableResource):
-        aws_credentials: AWSCredentialsResource
-        bucket_name: str
-
-    class EC2Resource(ConfigurableResource):
-        aws_credentials: AWSCredentialsResource
-
-    completed = {}
-
-    @asset
-    def my_asset(s3: S3Resource, ec2: EC2Resource):
-        assert s3.aws_credentials.username == "foo"
-        assert s3.aws_credentials.password == "bar"
-        assert s3.bucket_name == "my_bucket"
-
-        assert ec2.aws_credentials.username == "foo"
-        assert ec2.aws_credentials.password == "bar"
-
-        completed["yes"] = True
-
-    aws_credentials = AWSCredentialsResource.configure_at_launch()
-    defs = Definitions(
-        assets=[my_asset],
-        resources={
-            "aws_credentials": aws_credentials,
-            "s3": S3Resource(bucket_name="my_bucket", aws_credentials=aws_credentials),
-            "ec2": EC2Resource(aws_credentials=aws_credentials),
-        },
-    )
-
-    assert (
-        defs.get_implicit_global_asset_job_def()
-        .execute_in_process(
-            {
-                "resources": {
-                    "aws_credentials": {
-                        "config": {
-                            "username": "foo",
-                            "password": "bar",
-                        }
-                    }
-                }
-            }
-        )
-        .success
-    )
-    assert completed["yes"]
-
-
-def test_nested_resources_runtime_config_complex():
-    class CredentialsResource(ConfigurableResource):
-        username: str
-        password: str
-
-    class DBConfigResource(ConfigurableResource):
-        creds: CredentialsResource
-        host: str
-        database: str
-
-    class DBResource(ConfigurableResource):
-        config: DBConfigResource
-
-    completed = {}
-
-    @asset
-    def my_asset(db: DBResource):
-        assert db.config.creds.username == "foo"
-        assert db.config.creds.password == "bar"
-        assert db.config.host == "localhost"
-        assert db.config.database == "my_db"
-        completed["yes"] = True
-
-    credentials = CredentialsResource.configure_at_launch()
-    db_config = DBConfigResource.configure_at_launch(creds=credentials)
-    db = DBResource(config=db_config)
-
-    defs = Definitions(
-        assets=[my_asset],
-        resources={
-            "credentials": credentials,
-            "db_config": db_config,
-            "db": db,
-        },
-    )
-
-    assert (
-        defs.get_implicit_global_asset_job_def()
-        .execute_in_process(
-            {
-                "resources": {
-                    "credentials": {
-                        "config": {
-                            "username": "foo",
-                            "password": "bar",
-                        }
-                    },
-                    "db_config": {
-                        "config": {
-                            "host": "localhost",
-                            "database": "my_db",
-                        }
-                    },
-                }
-            }
-        )
-        .success
-    )
-    assert completed["yes"]
-
-    credentials = CredentialsResource.configure_at_launch()
-    db_config = DBConfigResource(creds=credentials, host="localhost", database="my_db")
-    db = DBResource(config=db_config)
-
-    defs = Definitions(
-        assets=[my_asset],
-        resources={
-            "credentials": credentials,
-            "db": db,
-        },
-    )
-
-    assert (
-        defs.get_implicit_global_asset_job_def()
-        .execute_in_process(
-            {
-                "resources": {
-                    "credentials": {
-                        "config": {
-                            "username": "foo",
-                            "password": "bar",
-                        }
-                    },
-                }
-            }
-        )
-        .success
-    )
-    assert completed["yes"]
-
-
-def test_enum_nested_resource_no_run_config() -> None:
-    class MyEnum(enum.Enum):
-        A = "a_value"
-        B = "b_value"
-
-    class ResourceWithEnum(ConfigurableResource):
-        my_enum: MyEnum
-
-    class OuterResourceWithResourceWithEnum(ConfigurableResource):
-        resource_with_enum: ResourceWithEnum
-
-    @asset
-    def asset_with_outer_resource(outer_resource: OuterResourceWithResourceWithEnum):
-        return outer_resource.resource_with_enum.my_enum.value
-
-    defs = Definitions(
-        assets=[asset_with_outer_resource],
-        resources={
-            "outer_resource": OuterResourceWithResourceWithEnum(
-                resource_with_enum=ResourceWithEnum(my_enum=MyEnum.A)
-            )
-        },
-    )
-
-    a_job = defs.get_implicit_global_asset_job_def()
-
-    result = a_job.execute_in_process()
-    assert result.success
-    assert result.output_for_node("asset_with_outer_resource") == "a_value"
-
-
-def test_enum_nested_resource_run_config_override() -> None:
-    class MyEnum(enum.Enum):
-        A = "a_value"
-        B = "b_value"
-
-    class ResourceWithEnum(ConfigurableResource):
-        my_enum: MyEnum
-
-    class OuterResourceWithResourceWithEnum(ConfigurableResource):
-        resource_with_enum: ResourceWithEnum
-
-    @asset
-    def asset_with_outer_resource(outer_resource: OuterResourceWithResourceWithEnum):
-        return outer_resource.resource_with_enum.my_enum.value
-
-    resource_with_enum = ResourceWithEnum.configure_at_launch()
-    defs = Definitions(
-        assets=[asset_with_outer_resource],
-        resources={
-            "resource_with_enum": resource_with_enum,
-            "outer_resource": OuterResourceWithResourceWithEnum(
-                resource_with_enum=resource_with_enum
-            ),
-        },
-    )
-
-    a_job = defs.get_implicit_global_asset_job_def()
-
-    # Case: I'm re-specifying the nested enum at runtime - expect the runtime config to override the resource config
-    result = a_job.execute_in_process(
-        run_config={"resources": {"resource_with_enum": {"config": {"my_enum": "B"}}}}
-    )
-    assert result.success
-    assert result.output_for_node("asset_with_outer_resource") == "b_value"
-
-
 def test_basic_enum_override_with_resource_instance() -> None:
     class BasicEnum(enum.Enum):
         A = "a_value"
@@ -872,243 +549,6 @@ def test_resources_which_return():
         .success
     )
     assert completed["yes"]
-
-
-def test_nested_function_resource():
-    out_txt = []
-
-    @resource
-    def writer_resource(context):
-        def output(text: str) -> None:
-            out_txt.append(text)
-
-        return output
-
-    class PostfixWriterResource(ConfigurableResourceFactory[Callable[[str], None]]):
-        writer: ResourceDependency[Callable[[str], None]]
-        postfix: str
-
-        def create_resource(self, context) -> Callable[[str], None]:
-            def output(text: str):
-                self.writer(f"{text}{self.postfix}")
-
-            return output
-
-    @asset
-    def my_asset(writer: ResourceParam[Callable[[str], None]]):
-        writer("foo")
-        writer("bar")
-
-    defs = Definitions(
-        assets=[my_asset],
-        resources={
-            "writer": PostfixWriterResource(writer=writer_resource, postfix="!"),
-        },
-    )
-
-    assert defs.get_implicit_global_asset_job_def().execute_in_process().success
-    assert out_txt == ["foo!", "bar!"]
-
-
-def test_nested_function_resource_configured():
-    out_txt = []
-
-    @resource(config_schema={"prefix": Field(str, default_value="")})
-    def writer_resource(context):
-        prefix = context.resource_config["prefix"]
-
-        def output(text: str) -> None:
-            out_txt.append(f"{prefix}{text}")
-
-        return output
-
-    class PostfixWriterResource(ConfigurableResourceFactory[Callable[[str], None]]):
-        writer: ResourceDependency[Callable[[str], None]]
-        postfix: str
-
-        def create_resource(self, context) -> Callable[[str], None]:
-            def output(text: str):
-                self.writer(f"{text}{self.postfix}")
-
-            return output
-
-    @asset
-    def my_asset(writer: ResourceParam[Callable[[str], None]]):
-        writer("foo")
-        writer("bar")
-
-    defs = Definitions(
-        assets=[my_asset],
-        resources={
-            "writer": PostfixWriterResource(writer=writer_resource, postfix="!"),
-        },
-    )
-
-    assert defs.get_implicit_global_asset_job_def().execute_in_process().success
-    assert out_txt == ["foo!", "bar!"]
-
-    out_txt.clear()
-
-    defs = Definitions(
-        assets=[my_asset],
-        resources={
-            "writer": PostfixWriterResource(
-                writer=writer_resource.configured({"prefix": "msg: "}), postfix="!"
-            ),
-        },
-    )
-
-    assert defs.get_implicit_global_asset_job_def().execute_in_process().success
-    assert out_txt == ["msg: foo!", "msg: bar!"]
-
-
-def test_nested_function_resource_runtime_config():
-    out_txt = []
-
-    @resource(config_schema={"prefix": str})
-    def writer_resource(context):
-        prefix = context.resource_config["prefix"]
-
-        def output(text: str) -> None:
-            out_txt.append(f"{prefix}{text}")
-
-        return output
-
-    class PostfixWriterResource(ConfigurableResourceFactory[Callable[[str], None]]):
-        writer: ResourceDependency[Callable[[str], None]]
-        postfix: str
-
-        def create_resource(self, context) -> Callable[[str], None]:
-            def output(text: str):
-                self.writer(f"{text}{self.postfix}")
-
-            return output
-
-    @asset
-    def my_asset(writer: ResourceParam[Callable[[str], None]]):
-        writer("foo")
-        writer("bar")
-
-    with pytest.raises(
-        CheckError,
-        match="Any partially configured, nested resources must be provided to Definitions",
-    ):
-        # errors b/c writer_resource is not configured
-        # and not provided as a top-level resource to Definitions
-        defs = Definitions(
-            assets=[my_asset],
-            resources={
-                "writer": PostfixWriterResource(writer=writer_resource, postfix="!"),
-            },
-        )
-
-    defs = Definitions(
-        assets=[my_asset],
-        resources={
-            "base_writer": writer_resource,
-            "writer": PostfixWriterResource(writer=writer_resource, postfix="!"),
-        },
-    )
-
-    assert (
-        defs.get_implicit_global_asset_job_def()
-        .execute_in_process(
-            {
-                "resources": {
-                    "base_writer": {
-                        "config": {
-                            "prefix": "msg: ",
-                        },
-                    },
-                },
-            }
-        )
-        .success
-    )
-    assert out_txt == ["msg: foo!", "msg: bar!"]
-
-
-def test_nested_config_class() -> None:
-    # Validate that we can nest Config classes in a pythonic resource
-
-    class User(Config):
-        name: str
-        age: int
-
-    class UsersResource(ConfigurableResource):
-        users: List[User]
-
-    executed = {}
-
-    @asset
-    def an_asset(users_resource: UsersResource):
-        assert len(users_resource.users) == 2
-        assert users_resource.users[0].name == "Bob"
-        assert users_resource.users[0].age == 25
-        assert users_resource.users[1].name == "Alice"
-        assert users_resource.users[1].age == 30
-
-        executed["yes"] = True
-
-    defs = Definitions(
-        assets=[an_asset],
-        resources={
-            "users_resource": UsersResource(
-                users=[
-                    User(name="Bob", age=25),
-                    User(name="Alice", age=30),
-                ]
-            )
-        },
-    )
-
-    assert defs.get_implicit_global_asset_job_def().execute_in_process().success
-    assert executed["yes"]
-
-
-def test_using_enum_simple() -> None:
-    executed = {}
-
-    class SimpleEnum(enum.Enum):
-        FOO = "foo"
-        BAR = "bar"
-
-    class MyResource(ConfigurableResource):
-        an_enum: SimpleEnum
-
-    @asset
-    def an_asset(my_resource: MyResource):
-        assert my_resource.an_enum == SimpleEnum.FOO
-        executed["yes"] = True
-
-    defs = Definitions(
-        assets=[an_asset],
-        resources={
-            "my_resource": MyResource(
-                an_enum=SimpleEnum.FOO,
-            )
-        },
-    )
-
-    assert defs.get_implicit_global_asset_job_def().execute_in_process().success
-    assert executed["yes"]
-    executed.clear()
-
-    defs = Definitions(
-        assets=[an_asset],
-        resources={
-            "my_resource": MyResource.configure_at_launch(),
-        },
-    )
-
-    assert (
-        defs.get_implicit_global_asset_job_def()
-        .execute_in_process(
-            {"resources": {"my_resource": {"config": {"an_enum": SimpleEnum.FOO.name}}}}
-        )
-        .success
-    )
-    assert executed["yes"]
 
 
 def test_using_enum_complex() -> None:
@@ -1555,3 +995,41 @@ def test_telemetry_dagster_resource():
             return True
 
     assert MyResource(my_value="foo")._is_dagster_maintained()  # noqa: SLF001
+
+
+def test_nested_config_class() -> None:
+    # Validate that we can nest Config classes in a pythonic resource
+
+    class User(Config):
+        name: str
+        age: int
+
+    class UsersResource(ConfigurableResource):
+        users: List[User]
+
+    executed = {}
+
+    @asset
+    def an_asset(users_resource: UsersResource):
+        assert len(users_resource.users) == 2
+        assert users_resource.users[0].name == "Bob"
+        assert users_resource.users[0].age == 25
+        assert users_resource.users[1].name == "Alice"
+        assert users_resource.users[1].age == 30
+
+        executed["yes"] = True
+
+    defs = Definitions(
+        assets=[an_asset],
+        resources={
+            "users_resource": UsersResource(
+                users=[
+                    User(name="Bob", age=25),
+                    User(name="Alice", age=30),
+                ]
+            )
+        },
+    )
+
+    assert defs.get_implicit_global_asset_job_def().execute_in_process().success
+    assert executed["yes"]
